feat: implementado service e controller refactor: alterei todos os id que possuem tipo Long para Integer

+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/curso/Curso.java
@@ -18,7 +18,7 @@ public class Curso {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_curso")
-    private Long idCurso;
+    private Integer idCurso;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "id_instituicao", nullable = false)
