+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/disciplina/DisciplinaRequestDTO.java
@@ -6,8 +6,8 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
 public record DisciplinaRequestDTO(
-        Long idCurso, // Pode ser null
-        @NotNull Long idTipoDisciplina,
+        Integer idCurso,
+        @NotNull Integer idTipoDisciplina,
         @NotBlank @Size(max = 10) String sigla,
         @NotBlank @Size(max = 150) String descricao,
         @NotNull @Min(1) Integer periodo,
