+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/instituicao/InstituicaoRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InstituicaoRepository extends JpaRepository<Instituicao, Long> {
+public interface InstituicaoRepository extends JpaRepository<Instituicao, Integer> {
 }
