+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/instituicao/InstituicaoResponseDTO.java
@@ -1,7 +1,7 @@
 package com.github.luisflp_dev.rest_escola.academico.instituicao;
 
 public record InstituicaoResponseDTO(
-        Long idInstituicao,
+        Integer idInstituicao,
         String sigla,
         String descricao
 ) {}
