+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/tipocurso/TipoCursoResponseDTO.java
@@ -1,6 +1,6 @@
 package com.github.luisflp_dev.rest_escola.academico.tipocurso;
 
 public record TipoCursoResponseDTO(
-        Long idTipoCurso,
+        Integer idTipoCurso,
         String descricao
 ) {}
