+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/tipodisciplina/TipoDisciplinaResponseDTO.java
@@ -1,6 +1,6 @@
 package com.github.luisflp_dev.rest_escola.academico.tipodisciplina;
 
 public record TipoDisciplinaResponseDTO(
-        Long idTipoDisciplina,
+        Integer idTipoDisciplina,
         String descricao
 ) {}
