+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/titulo/TituloRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TituloRepository extends JpaRepository<Titulo, Long> {
+public interface TituloRepository extends JpaRepository<Titulo, Integer> {
 }
