+++ src/main/java/com/github/luisflp_dev/rest_escola/academico/titulo/TituloResponseDTO.java
@@ -1,6 +1,6 @@
 package com.github.luisflp_dev.rest_escola.academico.titulo;
 
 public record TituloResponseDTO(
-        Long idTitulo,
+        Integer idTitulo,
         String descricao
 ){}
